add stylesheet for productlistentry
